+++ src/components/orders-payments/index.ts
@@ -1,5 +1,5 @@
 export { default as CheckoutHeader } from './CheckoutHeader';
-export { default as DeliveryAddress } from './DeliveryAddress';
+export { default as ShippingAddressForm } from './ShippingAddressForm';
 export { default as PaymentMethodSelector } from './PaymentMethodSelector';
 export { default as PaymentSummary } from './PaymentSummary';
 export * from './OrderProductList';
